+++ app/users.py
@@ -45,6 +45,8 @@ def find_existing_user(**kwargs) -> models.User:
     An instance of User that matches the UID."""
     try:
         user_uid = kwargs.get("user_uid", None)
+        if user_uid == None:
+            return None
 
         user_q = db.session.query(models.User)
         if user_uid: