+++ unittests/test_socket.py
@@ -22,9 +22,9 @@ class TestSocket(BaseMockLoginCase):
             username = "alden", vehicle = "1994 Toyota Supra")
         db.session.flush()
         messages_l = [
-            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -25.813579, longitude = 28.222248, rotation = 180.0, speed = 70.0, logged_at = time.time() * 1000),
-            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, rotation = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000),
-            dict(latitude = -25.813579, longitude = 28.222248, rotation = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+10000)
+            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -25.813579, longitude = 28.222248, bearing = 180.0, speed = 70.0, logged_at = time.time() * 1000),
+            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, bearing = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000),
+            dict(latitude = -25.813579, longitude = 28.222248, bearing = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+10000)
         ]
         # Authenticate the User.
         with self.app.test_client(user = aldos) as client:
@@ -56,9 +56,9 @@ class TestSocket(BaseMockLoginCase):
             username = "alden", vehicle = "1994 Toyota Supra")
         db.session.flush()
         messages_l = [
-            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -25.813579, longitude = 28.222248, rotation = 180.0, speed = 70.0, logged_at = time.time() * 1000),
-            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, rotation = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000),
-            dict(latitude = -25.813579, longitude = 28.222248, rotation = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+10000)
+            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -25.813579, longitude = 28.222248, bearing = 180.0, speed = 70.0, logged_at = time.time() * 1000),
+            dict(device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, bearing = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000),
+            dict(latitude = -25.813579, longitude = 28.222248, bearing = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+10000)
         ]
         # Authenticate the User.
         with self.app.test_client(user = aldos) as client:
@@ -82,7 +82,7 @@ class TestSocket(BaseMockLoginCase):
         with self.app.test_client(user = aldos) as client:
             # Launch a socket to join the world.
             join_world_d = dict(
-                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, rotation = 180.0, speed = 70.0, logged_at = time.time() * 1000)
+                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, bearing = 180.0, speed = 70.0, logged_at = time.time() * 1000)
             socket = socketio.test_client(self.app,
                 flask_test_client = client, auth = join_world_d)
             # Ensure the socket is connected.
@@ -106,7 +106,7 @@ class TestSocket(BaseMockLoginCase):
         db.session.flush()
         # Create a join world dictionary.
         join_world_d = world.RequestConnectAuthentication(
-                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, rotation = 180.0, speed = 70.0, logged_at = time.time() * 1000)
+                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, bearing = 180.0, speed = 70.0, logged_at = time.time() * 1000)
         # Create a dirty UserPlayer instance for this User and add it to session.
         _, dirty_user_player = self.make_user_player(aldos, join_world_d)
         db.session.add(dirty_user_player)
@@ -117,7 +117,7 @@ class TestSocket(BaseMockLoginCase):
         with self.app.test_client(user = aldos) as client:
             # Launch a socket to join the world with a new join world dict, but same FID etc.
             join_world_d = dict(
-                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, rotation = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000)
+                device_fid = uuid.uuid4().hex.lower(), latitude = -37.782737, longitude = 145.013383, bearing = 180.0, speed = 70.0, logged_at = (time.time() * 1000)+5000)
             socket = socketio.test_client(self.app,
                 flask_test_client = client, auth = join_world_d)
             # Ensure the socket is connected.
@@ -141,7 +141,7 @@ class TestSocket(BaseMockLoginCase):
             # Create a request to join the world, at the location of the track's very first point.
             track_start_pt = track.geodetic_point
             request_connect_auth_d = dict(                
-                device_fid = uuid.uuid4().hex.lower(), latitude = track_start_pt.y, longitude = track_start_pt.x, rotation = 180.0, speed = 70.0, logged_at = time_now)
+                device_fid = uuid.uuid4().hex.lower(), latitude = track_start_pt.y, longitude = track_start_pt.x, bearing = 180.0, speed = 70.0, logged_at = time_now)
             # Launch a socket to join the world.
             socket = socketio.test_client(self.app,
                 flask_test_client = client, auth = request_connect_auth_d)
@@ -153,10 +153,10 @@ class TestSocket(BaseMockLoginCase):
             # Now, send a request to start a race.
             # Start with the countdown position.
             countdown_d = dict(                
-                latitude = track_start_pt.y, longitude = track_start_pt.x, rotation = 180.0, speed = 70.0, logged_at = time_now+1000)
+                latitude = track_start_pt.y, longitude = track_start_pt.x, bearing = 180.0, speed = 70.0, logged_at = time_now+1000)
             # Make the started position.
             started_d = dict(                
-                latitude = track_start_pt.y, longitude = track_start_pt.x, rotation = 180.0, speed = 70.0, logged_at = time_now+5000)
+                latitude = track_start_pt.y, longitude = track_start_pt.x, bearing = 180.0, speed = 70.0, logged_at = time_now+5000)
             start_race_d = dict(
                 track_uid = track.uid, vehicle_uid = vehicle.uid, countdown_position = countdown_d, started_position = started_d)
             # Emit this event to the server.